Copia Seguriadad, Correcion de Contexto

+++ DAL/DAO/UsuarioDAO.cs
@@ -13,7 +13,7 @@ namespace DAL.DAO
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int id_usuario { get; set; }
+        public long id_usuario { get; set; }
         public string dni { get; set; }
         public string email { get; set; }
         public DateTime? expiracion_token { get; set; }
